:recycle: (Venue) Extract venue banner component
